requirements.txt
flask==3.0.0
starlink-grpc-core>=1.2.5
requests>=2.31.0
